update script to handle input files

+++ populate_s3.sh
@@ -1,16 +1,26 @@
-#!/bin/bash
+# Convert the provided files to HTML and push them to S3.
+# If no files are provided, then delete the S3 buckets contents
+# and convert all files
 
+#!/bin/bash
 AWS_S3_BUCKET="s3://braindump-bucket"
 
-aws s3 rm "$AWS_S3_BUCKET"
-git ls-files | while read -r file; do
+if [ $# -eq 0 ]; then
+    FILES=$(git ls-files)
+    aws s3 rm "$AWS_S3_BUCKET"
+else
+    FILES=$(echo "$@" | sed 's/ /\n/g')
+fi
+
+echo "$FILES" | while read -r file; do
+    [ ! -f "$file" ] && continue
 
     if [[ "$file" == *.org ]]; then
+        echo "converting $file"
         HTML_FILE=${file%.org}.html
         pandoc "$file" -o "$HTML_FILE" --wrap=none
         file="$HTML_FILE"
     fi
-
     aws s3 cp "$file" "$AWS_S3_BUCKET/$file"
 done
